# Refactor/attention