--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "TransformersLite"
 uuid = "6579f8b0-5af2-4c97-8a45-e81aa57d569b"
 authors = ["Lior Sinai <sinailior@gmail.com>"]
-version = "0.3.0"
+version = "0.3.1"
 
 [deps]
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"