Refactoring for the attention.
- Improved efficiency for the attention operation. Reshaping for 4D arrays is done in `scaled_dot_attention` instead of `batched_mul`. This reduces the number of reshapes from 6 (3x2 `batched_mul` calls) to 4.
- Renamed input variable from `dim_model` to `dim_in`.
- Now require the exact same input type for all input arrays. Technically this is a breaking change, albeit a minor one.