@LiorSinai 2024-03-29
Timing test for a matrix with 1,638,400 values:
```julia
X = randn(64, 100, 8, 32);
```

Main:
```julia
@time scaled_dot_attention(X, X, X);
  0.096476 seconds (28 allocations: 83.790 MiB, 5.48% gc time)
```

After refactor:
```julia
@time scaled_dot_attention(X, X, X);
  0.100125 seconds (20 allocations: 83.790 MiB, 12.89% gc time)
```

Allocations are reduced from 28 to 20.

Using BenchmarkTools
Main:
```julia
@benchmark scaled_dot_attention(X, X, X)
BenchmarkTools.Trial: 32 samples with 1 evaluation.
 Range (min … max):   91.470 ms … 316.651 ms  ┊ GC (min … max): 5.25% … 3.21%
 Time  (median):     154.740 ms               ┊ GC (median):    6.05%
 Time  (mean ± σ):   156.210 ms ±  49.134 ms  ┊ GC (mean ± σ):  5.99% ± 2.11%

        ▂█          ▅
  ▅█▁▁▅▁██▁▅▁▅▁▁▁▅▅▅█▁▁▅▁▁▁█▅▅▅▅▅▁▁▁▅▁▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅ ▁
  91.5 ms          Histogram: frequency by time          317 ms <

 Memory estimate: 83.79 MiB, allocs estimate: 28.
```

Refactor:
```julia
@benchmark scaled_dot_attention(X, X, X)
BenchmarkTools.Trial: 55 samples with 1 evaluation.
 Range (min … max):  82.326 ms … 109.700 ms  ┊ GC (min … max): 1.75% … 6.22%
 Time  (median):     88.738 ms               ┊ GC (median):    6.73%
 Time  (mean ± σ):   91.839 ms ±   6.329 ms  ┊ GC (mean ± σ):  6.70% ± 1.62%

          ▁ █ ▃▁▆▃           ▁
  ▄▁▁▁▁▁▁▁█▄█▄████▄▄▄▁▄▇▄▄▁▄▁█▁▁▇▁▁▁▄▁▄▄▄▁▁▁▇▁▁▄▄▁▁▁▁▄▁▁▄▁▁▁▄▄ ▁
  82.3 ms         Histogram: frequency by time          107 ms <

 Memory estimate: 83.79 MiB, allocs estimate: 20.
```

There is a definite decrease from about 150ms to 90ms.